Added TensorFlow specific directory
TensorFlow directory includes basic start of neural network code in TensorFlow. It currently does not work and has bugs that break it. 

While it is a decent start the bug with tf.squeeze and potentially the inability to use it with .map on our data set could mean bad things. Currently the shape of our audio data, the shape of an individual .wav element, consists of three fields (batch, samples, channels). tf.squeeze should be used to remove channels so we end up with element of shape (batch, samples). However tf.squeeze breaks, and removing tf.squeeze breaks the data display later on as we have tooooo many channels. 

I am stumped for now despite spending a few hours reading on this topic and will need help.